Finished adding support for the results page
The base results election page is completed, now users who view
the elections site after the election, or who log in can now
click a link to view the election results.

At the moment the election results just contain a table with
the position and the name and username of the person who was
elected to that position.